requirements.txt
opencv-python-headless>=4.5.0
numpy>=1.21.0
Pillow>=9.0.0
Flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0

